Move Renderer back to scene/
- This maintains a strict DAG of directories, which can be turned into libraries later: util <= vk <= glfw <= dear_imgui <= scene <= engine, editor
This DAG is currently loosely reflected in the order of sources in CMakeLists.txt.
