+++ facade-lib/CMakeLists.txt
@@ -86,9 +86,6 @@ target_sources(${PROJECT_NAME} PRIVATE
   include/facade/util/unique.hpp
   include/facade/util/visitor.hpp
 
-  include/facade/glfw/glfw.hpp
-  include/facade/glfw/input.hpp
-
   include/facade/vk/buffer.hpp
   include/facade/vk/cmd.hpp
   include/facade/vk/defer.hpp
@@ -110,10 +107,10 @@ target_sources(${PROJECT_NAME} PRIVATE
   include/facade/vk/vma.hpp
   include/facade/vk/wsi.hpp
 
-  include/facade/dear_imgui/dear_imgui.hpp
+  include/facade/glfw/glfw.hpp
+  include/facade/glfw/input.hpp
 
-  include/facade/engine/engine.hpp
-  include/facade/engine/renderer.hpp
+  include/facade/dear_imgui/dear_imgui.hpp
 
   include/facade/scene/camera.hpp
   include/facade/scene/fly_cam.hpp
@@ -122,8 +119,11 @@ target_sources(${PROJECT_NAME} PRIVATE
   include/facade/scene/material.hpp
   include/facade/scene/node.hpp
   include/facade/scene/rect.hpp
+  include/facade/scene/renderer.hpp
   include/facade/scene/scene.hpp
   include/facade/scene/transform.hpp
+
+  include/facade/engine/engine.hpp
 
   include/facade/editor/common.hpp
   include/facade/editor/inspector.hpp
@@ -134,8 +134,6 @@ target_sources(${PROJECT_NAME} PRIVATE
   src/util/image.cpp
   src/util/logger.cpp
   src/util/time.cpp
-
-  src/glfw/glfw.cpp
 
   src/vk/buffer.cpp
   src/vk/cmd.cpp
@@ -155,18 +153,20 @@ target_sources(${PROJECT_NAME} PRIVATE
   src/vk/vk.cpp
   src/vk/vma.cpp
 
+  src/glfw/glfw.cpp
+
   src/dear_imgui/dear_imgui.cpp
 
   src/detail/gltf.cpp
   src/detail/gltf.hpp
 
-  src/engine/engine.cpp
-  src/engine/renderer.cpp
-
   src/scene/camera.cpp
   src/scene/material.cpp
+  src/scene/renderer.cpp
   src/scene/scene.cpp
   src/scene/transform.cpp
+
+  src/engine/engine.cpp
 
   src/editor/common.cpp
   src/editor/inspector.cpp
