+++ facade-lib/src/engine/engine.cpp
@@ -1,5 +1,5 @@
 #include <facade/engine/engine.hpp>
-#include <facade/engine/renderer.hpp>
+#include <facade/scene/renderer.hpp>
 #include <facade/scene/scene.hpp>
 #include <facade/vk/vk.hpp>
