+++ facade-lib/src/scene/renderer.cpp
@@ -1,5 +1,5 @@
 #include <facade/dear_imgui/dear_imgui.hpp>
-#include <facade/engine/renderer.hpp>
+#include <facade/scene/renderer.hpp>
 #include <facade/util/error.hpp>
 #include <facade/util/logger.hpp>
 #include <facade/util/time.hpp>
